Use typed proxy for ipcRenderer.
Similar to the previous commit, we now use a typed proxy for ipcRenderer to
ensure we don't introduce typos in event names on the Vue side.  This
requires importint ipcRenderer from our wrapper script rather than
importing it from electron directly.
