+++ pkg/rancher-desktop/pages/Dialog.vue
@@ -2,8 +2,9 @@
 import os from 'os';
 
 import { Checkbox } from '@rancher/components';
-import { ipcRenderer } from 'electron';
 import Vue from 'vue';
+
+import { ipcRenderer } from '@pkg/utils/ipcRenderer';
 
 export default Vue.extend({
   name:       'rd-dialog',
