+++ pkg/rancher-desktop/store/applicationSettings.ts
@@ -1,9 +1,9 @@
-import { ipcRenderer } from 'electron';
 
 import { ActionContext, MutationsType } from './ts-helpers';
 
 import { load as loadSettings } from '@pkg/config/settings';
 import type { PathManagementStrategy } from '@pkg/integrations/pathManager';
+import { ipcRenderer } from '@pkg/utils/ipcRenderer';
 
 /**
  * State is the type of the state we are maintaining in this store.
