+++ pkg/rancher-desktop/store/preferences.ts
@@ -1,10 +1,10 @@
-import { ipcRenderer } from 'electron';
 import _ from 'lodash';
 
 import { ActionContext, MutationsType } from './ts-helpers';
 
 import { defaultSettings, Settings } from '@pkg/config/settings';
 import type { ServerState } from '@pkg/main/commandServer/httpCommandServer';
+import { ipcRenderer } from '@pkg/utils/ipcRenderer';
 import { RecursiveKeys, RecursivePartial, RecursiveTypes } from '@pkg/utils/typeUtils';
 
 import type { GetterTree } from 'vuex';
