+++ pkg/rancher-desktop/typings/electron-ipc.d.ts
@@ -4,8 +4,8 @@
 
 import Electron from 'electron';
 
-import { ServiceEntry } from '@pkg/backend/k8s';
-import { RecursivePartial } from '@pkg/utils/typeUtils';
+import type { ServiceEntry } from '@pkg/backend/k8s';
+import type { RecursivePartial } from '@pkg/utils/typeUtils';
 /**
  * IpcMainEvents describes events the renderer can send to the main process,
  * i.e. ipcRenderer.send() -> ipcMain.on().
@@ -46,6 +46,9 @@ export interface IpcMainEvents {
   // #region dialog
   'dialog/load': () => void;
   'dialog/ready': () => void;
+  'dialog/mounted': () => void;
+  /** For message box only */
+  'dialog/close': (...args: any[]) => void;
   // #endregion
 
   // #region sudo-prompt
@@ -71,6 +74,9 @@ export interface IpcMainEvents {
   'dashboard-close': () => void;
 
   'diagnostics/run': () => void;
+
+  /** Only for the preferences window */
+  'preferences/load': () => void;
 }
 
 /**
@@ -82,10 +88,10 @@ export interface IpcMainInvokeEvents {
   'transient-settings-fetch': () => import('@pkg/config/transientSettings').TransientSettings;
   'transient-settings-update': (arg: RecursivePartial<import('@pkg/config/transientSettings').TransientSettings>) => void;
   'service-fetch': (namespace?: string) => import('@pkg/backend/k8s').ServiceEntry[];
-  'service-forward': (service: { namespace: string, name: string, port: string | number, listenPort?: number }, state: boolean) => void;
+  'service-forward': (service: ServiceEntry, state: boolean) => void;
   'get-app-version': () => string;
   'show-message-box': (options: Electron.MessageBoxOptions) => Promise<Electron.MessageBoxReturnValue>;
-  'api-get-credentials': () => { user: string, password: string, port: number };
+  'api-get-credentials': () => { user: string, password: string, port: number, pid: number };
 
   // #region main/imageEvents
   'images-mounted': (mounted: boolean) => {imageName: string, tag: string, imageID: string, size: string}[];
@@ -126,70 +132,13 @@ export interface IpcRendererEvents {
   // #endregion
 
   // #region dialog
-  'dalog/populate': (...args: any) => void;
+  'dialog/populate': (...args: any) => void;
   'dialog/size': (size: {width: number, height: number}) => void;
   'dialog/options': (...args: any) => void;
-  'dialog/close': (...args: any) => void;
   'dashboard-open': () => void;
   // #endregion
 
   // #region api
   'api-credentials': (credentials: {user: string, password: string, port: number}) => void;
   // #endregion
-}
-
-declare module 'electron' {
-  // We mark the default signatures as deprecated, so that ESLint will throw an
-  // error if they are used.
-
-  interface IpcRenderer {
-    on<eventName extends keyof IpcRendererEvents>(
-      channel: eventName,
-      listener: (event: Electron.IpcRendererEvent, ...args: globalThis.Parameters<IpcRendererEvents[eventName]>) => void
-    ): this;
-    /** @deprecated */
-    on(channel: string, listener: (...args: any[]) => void): this;
-
-    once<eventName extends keyof IpcRendererEvents>(
-      channel: eventName,
-      listener: (event: Electron.IpcRendererEvent, ...args: globalThis.Parameters<IpcRendererEvents[eventName]>) => void
-    ): this;
-    /** @deprecated */
-    once(channel: string, listener: (...args: any[]) => void): this;
-
-    removeListener<eventName extends keyof IpcRendererEvents>(
-      channel: eventName,
-      listener: (event: Electron.IpcRendererEvent, ...args: globalThis.Parameters<IpcRendererEvents[eventName]>) => void
-    ): this;
-    /** @deprecated */
-    removeListener(channel: string, listener: (...args: any[]) => void): this;
-
-    removeAllListeners<eventName extends keyof IpcRendererEvents>(channel?: eventName): this;
-    /** @deprecated */
-    removeAllListeners(channel: string): this;
-
-    // When the renderer side is implement in JavaScript (rather than TypeScript),
-    // the type checking for arguments seems to fail and always prefers the
-    // generic overload (which we want to avoid) rather than the specific overload
-    // we provide here.  Until we convert all of the Vue components to TypeScript,
-    // for now we will need to forego checking the arguments.
-    send<eventName extends keyof IpcMainEvents>(
-      channel: eventName,
-      ...args: any[],
-    ): this;
-    /** @deprecated */
-    send(channel: string, ...args: any[]): void;
-
-    // When the renderer side is implement in JavaScript (rather than TypeScript),
-    // the type checking for arguments seems to fail and always prefers the
-    // generic overload (which we want to avoid) rather than the specific overload
-    // we provide here.  Until we convert all of the Vue components to TypeScript,
-    // for now we will need to forego checking the arguments.
-    invoke<eventName extends keyof IpcMainInvokeEvents>(
-      channel: eventName,
-      ...args: any[],
-    ): Promise<ReturnType<IpcMainInvokeEvents[eventName]>>;
-    /** @deprecated */
-    invoke(channel: string, ...args: any[]): any;
-  }
 }
